// src/constants/message-socket-enum.ts
export enum MessageSocketEnum {
  ALL_ACTIVE_GAMES = 'all_active_games',
  CONNECT_TO_GAME = 'connect_to_game',
  CREATE_GAME = 'create_game',
  CREATE_PLAYER = 'create_player',
  GAME_OVER = 'game_over',
  MY_SHOT = 'my_shot',
  OPPONENT_SHOT = 'opponent_shot',
  PUNCH = 'punch',
  START_GAME = 'start_game',
  STEP = 'step'
}